# api/docs/swagger.yaml
basePath: /
definitions:
  signin.SigninRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  signin.SigninResponse:
    properties:
      token:
        type: string
    type: object
  signup.SignupRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8181
info:
  contact: {}
  description: API de base pour jeu en Go
  title: Gowebgame API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Affiche un message de bienvenue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accueil
      tags:
      - Accueil
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authentifie un utilisateur existant et renvoie un token JWT
      parameters:
      - description: Identifiants de l'utilisateur
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/signin.SigninRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT généré
          schema:
            $ref: '#/definitions/signin.SigninResponse'
        "400":
          description: Requête invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Identifiants incorrects
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Connexion utilisateur
      tags:
      - Authentification
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Crée un nouveau compte utilisateur avec un identifiant unique et
        un mot de passe
      parameters:
      - description: Informations de l'utilisateur
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/signup.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message de confirmation
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Erreur de validation
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur serveur ou utilisateur déjà existant
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Inscription utilisateur
      tags:
      - Authentification
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" suivi d'un espace et de votre token JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"